move crypto into pkg
